refactor: Move helpers functions to separate file and made one parameterized test
